feat: changed the input field to text area
